fix: Scores aren't guaranteed to be in order by date
This can happen, for example, when a new language is added. The current logic for populating the language data starts from the most current date and moves backward. I suppose we could change that logic, but it seems better to make the frontend less brittle, and thankfully there doesn't seem to be any performance hit.